NICE-178 Move components to News dir to declutter main dir
