+++ web/src/pages/news/articles/index.page.tsx
@@ -9,7 +9,7 @@ import { PageHeader } from "@nice-digital/nds-page-header";
 
 import { FeaturedStory } from "@/components/Storyblok/News/FeaturedStory/FeaturedStory";
 import { NewsList } from "@/components/Storyblok/News/NewsList/NewsList";
-import { NewsListNav } from "@/components/Storyblok/NewsListNav/NewsListNav";
+import { NewsListNav } from "@/components/Storyblok/News/NewsListNav/NewsListNav";
 import { NewsListPagination } from "@/components/Storyblok/NewsListPagination/NewsListPagination";
 import { logger } from "@/logger";
 import { NewsStory } from "@/types/News";
